feat(6800): extract block pipeline transport fed tests
