+++ crates/kamn-core/tests/block_pipeline_transport_fed/transport_event_feed_contract_tests.rs
@@ -1,0 +1,4 @@
+#[path = "transport_event_feed_contract_tests/event_feed_decode_contract_tests.rs"]
+mod event_feed_decode_contract_tests;
+#[path = "transport_event_feed_contract_tests/canonical_payload_contract_tests.rs"]
+mod canonical_payload_contract_tests;
