+++ crates/kamn-core/tests/block_pipeline_transport_fed/transport_pipeline_contract_tests.rs
@@ -1,0 +1,4 @@
+#[path = "transport_pipeline_contract_tests/pipeline_round_contract_tests.rs"]
+mod pipeline_round_contract_tests;
+#[path = "transport_pipeline_contract_tests/fork_choice_contract_tests.rs"]
+mod fork_choice_contract_tests;
